+++ Mods/lx_enhanced_divine_combat_3ff156e2-289e-4dac-81f5-a44e3e304163/Story/RawFiles/Lua/BootstrapServer.lua
@@ -6,6 +6,7 @@ Ext.Require("LXDGM_Potions.lua")
 Ext.Require("LXDGM_Talents.lua")
 Ext.Require("LXDGM_Weapons.lua")
 Ext.Require("LXDGM_StatsPatching.lua")
+Ext.Require("LXDGM_CustomBonuses.lua")
 
 ---- General Functions ----
 ---@param character EsvCharacter
@@ -62,9 +63,9 @@ function CheckStatChange(character)
 			local stored = GetVarInteger(character, "LX_Check_"..attr)
 			local storedBase = GetVarInteger(character, "LX_Check_Base_"..attr)
 			if stat ~= stored then
-				--print(attr.." changed from "..stored.." to "..stat)
+				Ext.Print(attr.." changed from "..stored.." to "..stat)
 				changed = true
-				--print("Stat change for "..character)
+				Ext.Print("Stat change for "..character)
 				SetVarInteger(character, "LX_Check_Base_"..attr, baseStat)
 				SetVarInteger(character, "LX_Changed_Base_"..attr, baseStat - storedBase)
 				SetVarInteger(character, "LX_Check_"..attr, stat)
@@ -106,9 +107,10 @@ function ApplyOverhaulAttributeBonuses(character)
 
 	---- Attribute Bonus
 	-- Movement Bonus
-	ApplyBonusOnce(character, "Finesse", "Movement", Ext.ExtraData.DGM_FinesseMovementBonus, false)
-	-- Accuracy bonus
-	ApplyBonusOnce(character, "Intelligence", "Accuracy", Ext.ExtraData.DGM_IntelligenceAccuracyBonus, false)
+	-- ApplyBonusOnce(character, "Finesse", "Movement", Ext.ExtraData.DGM_FinesseMovementBonus, false)
+	-- ApplyBonusOnce(character, "Finesse", "CriticalChance", Ext.ExtraData.DGM_FinesseCritChance, false)
+	-- -- Accuracy bonus
+	-- ApplyBonusOnce(character, "Intelligence", "Accuracy", Ext.ExtraData.DGM_IntelligenceAccuracyBonus, false)
 	-- Memory Bonus for Mnemonic
 	if CharacterHasTalent(character, "Memory") == 1 then 
 		ApplyBonusOnce(character, "Memory", "Memory", 1, true) 
@@ -145,7 +147,7 @@ end
 ---@param character EsvCharacter
 function ApplyOverhaulBonuses(character)
 	ApplyOverhaulAttributeBonuses(character)
-	ApplyOverhaulWeaponAbilityBonuses(character)
+	--ApplyOverhaulWeaponAbilityBonuses(character)
 end
 
 ---@param character EsvCharacter
@@ -180,6 +182,34 @@ local function RequestClientID(call,id,callbackID)
 end
 
 Ext.RegisterNetListener("PDGM_RequestClientID", RequestClientID)
+
+local function CheckStatChangeNetID(message, netID)
+	local char = Ext.GetCharacter(tonumber(netID))
+	SetStoryEvent(char.MyGuid, "DGM_SyncAttributeBonuses")
+end
+
+Ext.RegisterNetListener("DGM_UpdateCharacter", CheckStatChangeNetID)
+
+local function CleanBoosts(char)
+	print("Clearing boosts...")
+	TimerLaunch("DGM_GriffToAtusaFix", 50)
+end
+
+local function DGM_consoleCmd(cmd, ...)
+	local params = {...}
+	for i=1,10,1 do
+		local par = params[i]
+		if par == nil then break end
+		if type(par) == "string" then
+			par = par:gsub("&", " ")
+			par = par:gsub("\\ ", "&")
+			params[i] = par
+		end
+	end
+    if cmd == "DGM_CleanPermaBoosts" then CleanBoosts() end
+end
+
+Ext.RegisterConsoleCommand("DGM_CleanPermaBoosts", DGM_consoleCmd)
 
 
 ---- Create calls and queries