+++ Mods/lx_enhanced_divine_combat_3ff156e2-289e-4dac-81f5-a44e3e304163/Story/RawFiles/Lua/LXDGM_Helpers.lua
@@ -45,20 +45,26 @@ function CharGetDGMAttributeBonus(char, next)
 	local strength = stats.Strength - Ext.ExtraData.AttributeBaseValue
 	local finesse = stats.Finesse - Ext.ExtraData.AttributeBaseValue
 	local intelligence = stats.Intelligence - Ext.ExtraData.AttributeBaseValue
+	local wits = stats.Wits - Ext.ExtraData.AttributeBaseValue
 
 	local bonus = {
 		str = math.floor(strength+next),
 		strGlobal = math.floor((strength+next) * Ext.ExtraData.DGM_StrengthGlobalBonus),
 		strWeapon = math.floor((strength+next) * Ext.ExtraData.DGM_StrengthWeaponBonus),
-		strDot = math.floor((strength+next) * Ext.ExtraData.DGM_StrengthDoTBonus),
+		strRes = math.floor((strength+next) * Ext.ExtraData.DGM_StrengthResistanceIgnore),
 		fin = math.floor(finesse+next),
 		finGlobal = math.floor((finesse+next) * Ext.ExtraData.DGM_FinesseGlobalBonus),
 		finDodge = round((finesse+next) * Ext.ExtraData.DodgingBoostFromAttribute * 100, 0),
 		finMovement = round((finesse+next) * Ext.ExtraData.DGM_FinesseMovementBonus / 100, 2),
+		finCrit = math.floor((finesse+next) * Ext.ExtraData.DGM_FinesseCritChance),
 		int = math.floor(intelligence+next),
 		intGlobal = math.floor((intelligence+next) * Ext.ExtraData.DGM_IntelligenceGlobalBonus),
 		intSkill = math.floor((intelligence+next) * Ext.ExtraData.DGM_IntelligenceSkillBonus),
 		intAcc = math.floor((intelligence+next) * Ext.ExtraData.DGM_IntelligenceAccuracyBonus),
+		wits = math.floor(wits+next),
+		witsCrit = math.floor(wits+next * Ext.ExtraData.CriticalBonusFromWits),
+		witsIni = math.floor(wits+next * Ext.ExtraData.InitiativeBonusFromWits),
+		witsDot = math.floor((wits+next) * Ext.ExtraData.DGM_WitsDotBonus),
 		dual = math.floor(Ext.ExtraData.CombatAbilityDamageBonus * (stats.DualWielding+next)),
 		dualDodge = math.floor(Ext.ExtraData.CombatAbilityDodgingBonus * (stats.DualWielding+next)),
 		dualOff = math.floor(Ext.ExtraData.DGM_DualWieldingOffhandBonus * (stats.DualWielding+next)),